+++ src/main/java/requious/recipe/RequirementActive.java
@@ -36,7 +36,7 @@ public class RequirementActive extends RequirementBase {
 
     @Override
     public ConsumptionResult createResult() {
-        return new ConsumptionResult.Integer(this, 0);
+        return new ConsumptionResult.Long(this, 0L);
     }
 
     @Override
