+++ src/main/java/requious/recipe/RequirementDuration.java
@@ -47,7 +47,7 @@ public class RequirementDuration extends RequirementBase {
 
     @Override
     public ConsumptionResult createResult() {
-        return new ConsumptionResult.Integer(this, 0);
+        return new ConsumptionResult.Long(this, 0L);
     }
 
     @Override
