+++ src/main/java/requious/recipe/RequirementJEI.java
@@ -34,7 +34,7 @@ public class RequirementJEI extends RequirementBase {
 
     @Override
     public ConsumptionResult createResult() {
-        return new ConsumptionResult.Integer(this, 0);
+        return new ConsumptionResult.Long(this, 0L);
     }
 
     @Override
